chore: update sample
