+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@orillusion/core",
-    "version": "0.7.3",
+    "version": "0.8.0-beta.1",
     "author": "Orillusion",
     "description": "Orillusion WebGPU Engine",
     "type": "module",
