+++ samples/lights/Sample_CSM.ts
@@ -11,8 +11,8 @@ class Sample_CSM {
     viewCamera: Camera3D;
     async run() {
         Engine3D.setting.shadow.autoUpdate = true;
-        Engine3D.setting.shadow.shadowSize = 1024;
-        Engine3D.setting.shadow.shadowBound = 400;
+        Engine3D.setting.shadow.shadowSize = 2048;
+        Engine3D.setting.shadow.shadowBound = 512;
         await Engine3D.init({ renderLoop: () => { this.loop(); } });
 
         GUIHelp.init();
