+++ samples/render/Sample_BlendMode.ts
@@ -55,19 +55,19 @@ class Sample_BlendMode2 {
             let plane = new Object3D();
             let renderer = plane.addComponent(MeshRenderer);
             let material = new UnLitMaterial();
-            material.baseMap = await Engine3D.res.loadTexture("particle/T_Fx_Object_229.png");;
+            material.baseMap = await Engine3D.res.loadTexture("particle/T_Fx_Object_229.png");
+            material.blendMode = BlendMode.NORMAL;
             renderer.material = material;
             renderer.geometry = new PlaneGeometry(100, 100, 1, 1);
-            material.blendMode = BlendMode.NORMAL;
             this.scene.addChild(plane);
 
             GUIHelp.init();
 
             // blend mode
-            let blendMode = {
-                NONE: BlendMode.NONE,
-                NORMAL: BlendMode.NORMAL,
-                ADD: BlendMode.ADD,
+            let blendMode = {}
+            for(let i in BlendMode){
+                if(!i.match(/\d/))
+                    blendMode[i] = BlendMode[i];
             }
             // change blend mode by click dropdown box
             GUIHelp.add({ blendMode: material.blendMode }, 'blendMode', blendMode).onChange((v) => {
